+++ src/frontend/ts/constants/chess.ts
@@ -1,5 +1,0 @@
-const ChessConstants = {
-  initialFen: 'rnbqkbnr/pppppppp/8/8/8/8/PPPPPPPP/RNBQKBNR'
-}
-
-export default ChessConstants